chore: rename LogPlugin to DevLogPlugin
