+++ lib/dev.js
@@ -13,7 +13,7 @@ module.exports = async function dev (sourceDir, cliOptions = {}) {
   const prepare = require('./prepare')
   const logger = require('./util/logger')
   const HeadPlugin = require('./webpack/HeadPlugin')
-  const LogPlugin = require('./webpack/LogPlugin')
+  const DevLogPlugin = require('./webpack/DevLogPlugin')
   const createClientConfig = require('./webpack/createClientConfig')
   const { applyUserWebpackConfig } = require('./util')
   const { frontmatterEmitter } = require('./webpack/markdownLoader')
@@ -24,7 +24,7 @@ module.exports = async function dev (sourceDir, cliOptions = {}) {
   // setup watchers to update options and dynamically generated files
   const update = () => {
     prepare(sourceDir).catch(err => {
-      console.error(chalk.red(err.stack))
+      console.error(logger.error(chalk.red(err.stack), false))
     })
   }
 
@@ -73,7 +73,7 @@ module.exports = async function dev (sourceDir, cliOptions = {}) {
 
   config
   .plugin('fancy-log')
-  .use(LogPlugin, [{
+  .use(DevLogPlugin, [{
     port,
     displayHost,
     publicPath: options.publicPath
